0.0.5
My json file, this file contains all the information that is going to be in the resume. The idea is that it will be posible to alter this file from the admin panel at first, and at some point it will be posible to alter it by a form in the admin panel.