0.1.9 - Admin resume panel almost done
In general is done the resume panel on the admin site, I did a lot of changes from my original idea, now it makes use of the database to store a log of changes of the resume so the user can rollback in any moment wanted (something like this, github).
Also the web reads the json from the database and not from the file, I need to make that if are no records in the database it generates one automatically from the file.