0.2.6 - The Posts in the admin part is ready
I think I have completed almost everything is needed for the control and edition of posts in the admin part, I will add later other things like a visible switch button directly on table.
But in general is ready to use.

My love :heart: to this complements I used:
https://select2.org/
https://summernote.org/

This was not my first time with them but they are still amazing and simplified a lot the work and design.