0.3.5 - Portfolio is ready
I finished Portfolio, now it is accessible to the user and the admin can upload projects there.
The only thing of general functions that is required is the contact form, I hope I can finish it tomorrow and from then focus in make better and finish what I let aside.